moved op_data_config to pulse_controller to consolidate simulation data into one place
